Replace Makefile with CMakeLists Revamp README file to include build instructions and hexcat examples hexcat.c, hexcat.h - Move function declaration to h file hextools.c, hextools.h - New files that contains implementations and declarations of unnibble() and unhex() hexecho.c - include hextools module .gitignore - Add for conveniece, ignoring the build/ dir
