+++ hexcat.h
@@ -1,0 +1,3 @@
+#include "hextools.h"
+
+int cat_fd(int ifd, int raw);
