+++ hexecho.c
@@ -7,9 +7,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <ctype.h>
-     
-unsigned char	unhex(unsigned char hex1, unsigned char hex2);
-int	unnibble(unsigned char x);
+#include "hextools.h"     
 
 int	main(int argc, char *argv[])
 {
@@ -82,25 +80,3 @@ int	main(int argc, char *argv[])
 	fflush(stdout);
 }
 
-
-unsigned char	unhex(unsigned char hex1, unsigned char hex2)
-{
-    return ( unnibble(hex1)*16+unnibble(hex2) );
-}
-
-int	unnibble(unsigned char x)
-{
-    if((x>='0') && (x<='9'))
-    {
-		return (x-'0');
-    }
-    else if((x>='a') && (x<='f'))
-    {
-		return (x-'a'+10);
-    }
-    else if((x>='A') && (x<='F'))
-    {
-		return(x-'A'+10);
-	}
-    return -1;
-}
