+++ hextools.h
@@ -1,0 +1,3 @@
+unsigned char unhex(unsigned char hex1, unsigned char hex2);
+
+int	unnibble(unsigned char x);
